% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotivate.R
\name{demotivate}
\alias{demotivate}
\title{Generate demotivational quotes for personal use.}
\usage{
demotivate(cat = "any", seed)
}
\arguments{
\item{cat}{a demotivational category to choose from. The function defaults to 'any'.}

\item{seed}{a numeric or character seed for reproducible results. The function defaults to 'NULL'.}
}
\value{
\code{demotivate()} takes a preferred category and an optional reproducible seed
to return a demotivational quote for the user to feel worse at any given time.
}
\description{
\code{demotivate()} is a tool to demotivate even the highly motivated individual.
Use with caution and remember to not take it personally.
}
\details{
When cat is 'any' (default), the function combines all category data frames and then
samples from the new,combined data frame.

There are a total of six categories: family, life, work, science, TV, and rednecks.
When a specific category is chosen the function will sample from a demotivational quote data
frame and output a random quote for your spiritual and mental decay.
}
\examples{
demotivate("any",seed = "lanky puppy")
demotivate("work")
}
\author{
JP Monteagudo
}
