% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despair_message.R
\name{despair.message}
\alias{despair.message}
\title{Display package version for \pkg{despair}}
\usage{
despair.message()
}
\value{
\code{despair.message()} returns a message about the install version
of \pkg{despair}.
}
\description{
\code{despair.message()} produces a message about the package version
and the version of R making use of this package.
}
\examples{
despair.message()


}
\author{
JP Monteagudo
}
