% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsm_5}
\alias{dsm_5}
\title{A data frame of DSM-5 adjectives}
\format{
a data frame with 150 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
dsm_5
}
\description{
This is a data frame of adjectives used in the DSM-5 to describe mental states
}
\keyword{datasets}
