% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{family}
\alias{family}
\title{A data frame of family-related demotivational quotes}
\format{
a data frame with 13 observations and1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
family
}
\description{
This is a data frame of family-related demotivational quotes in English taken from American popular culture.
}
\keyword{datasets}
