% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life}
\alias{life}
\title{A data frame of life-related demotivational quotes.}
\format{
a data frame with 27 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
life
}
\description{
This is a data frame if life-related demotivational quotes in English taken from American popular culture.
}
\keyword{datasets}
