% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lit}
\alias{lit}
\title{A data frame of modern and contemporary literature motivational quotes.}
\format{
a data frame with 22 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
lit
}
\description{
This is a data frame of modern and contemporary literature quotes to be used with \code{motivate()}
}
\keyword{datasets}
