% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modern}
\alias{modern}
\title{A data frame of modern and contemporary historical and pop-culture icons}
\format{
a data frame of with 40 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
modern
}
\description{
This is a data frame of modern and contemporary historical and pop-culture icons used in  \code{motivate()}
}
\keyword{datasets}
