% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivate.R
\name{motivate}
\alias{motivate}
\title{General motivation quotes for personal use}
\usage{
motivate(cat = "any", seed)
}
\arguments{
\item{cat}{a motivational category to choose from. The function defaults to "any".}

\item{seed}{a numeric or character seed for reproducible results. The function defaults to "NULL".}
}
\value{
\code{motivate()} takes a preferred category and an optional reproducible seed
to return a motivational quote for the user to feel better at any given time.
}
\description{
\code{motivate()} is a tool to lift up even the unmotivated individual.
Use with caution and remember— DO take it personally.
}
\details{
When cat is 'any' (default), the function combines all category data frames and then
samples from the new,combined data frame.

There are a total of five categories: psych, stoic, religion, lit(literature), and modern (history)
When a specific category is chosen the function will sample from a motivational quote data
frame and output a random quote for your spiritual and mental growth.
}
\examples{

motivate("any",seed = "lethargic raccoon")
motivate("lit")
motivate("stoic", seed = 2899)
}
\author{
JP Monteagudo
}
