% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psych}
\alias{psych}
\title{A data frame of psychology-related quotes}
\format{
a data frame with 34 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
psych
}
\description{
This is a data frame of psychology-related quotes to be used in \code{motivate()}
}
\keyword{datasets}
