% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rednecks}
\alias{rednecks}
\title{A data frame of sweet redneck humor}
\format{
a data frame with 21 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
rednecks
}
\description{
This is a data frame of sweet redneck humor taken from Southern popular culture.
}
\keyword{datasets}
