% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{religion}
\alias{religion}
\title{A data frame of religious motivational quotes}
\format{
a data frame with 45 observations and 1 column.
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
religion
}
\description{
This is a data frame of religiously-inspired motivational quotes to be used in \code{motivate()}
}
\keyword{datasets}
