% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{science}
\alias{science}
\title{A data frame of science-related demotivational quotes}
\format{
a data frame of 27 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
science
}
\description{
This is a data frame of science-related demotivational quotes.
}
\keyword{datasets}
