% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shake_adjectives}
\alias{shake_adjectives}
\title{A data frame of Shakespeare's adjectives}
\format{
a data frame with 100 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
shake_adjectives
}
\description{
This is a data frame of adjectives used in Shakespeare's works.
}
\keyword{datasets}
