% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shake_animals}
\alias{shake_animals}
\title{A data frame of Shakespeare's animals}
\format{
a data frame with 150 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
shake_animals
}
\description{
This is a data frame of animals used in Shakespeare's works
}
\keyword{datasets}
