% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shake_chars}
\alias{shake_chars}
\title{A data frame of Shakespeare's characters}
\format{
a data frame with 137 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
shake_chars
}
\description{
This is a data frame of Shakespeare's characters
}
\keyword{datasets}
