% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shake_clrs}
\alias{shake_clrs}
\title{A data frame of Shakespeare and unusual colors}
\format{
a data frame with 300 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
shake_clrs
}
\description{
This is a data frame of colors used in Shakespeare's work and other unusual colors
}
\keyword{datasets}
