% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shake_jobs}
\alias{shake_jobs}
\title{A data frame of Shakespeare's jobs}
\format{
a data frame with 103 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
shake_jobs
}
\description{
This is a data frame of professions used in Shakespeare's works.
}
\keyword{datasets}
