% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shake_things}
\alias{shake_things}
\title{A data frame of Shakespeare's objects}
\format{
a data frame with 150 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
shake_things
}
\description{
This is a data frame of objects used in Shakespeare's works.
}
\keyword{datasets}
