% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stoic}
\alias{stoic}
\title{A data frame o stoic motivational quotes}
\format{
a data frame with 32 observations and 1 column.
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
stoic
}
\description{
This is a data frame of stoic motivational quotes to be used in \code{motivate()}
}
\keyword{datasets}
