% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tv}
\alias{tv}
\title{A data frame of tv and film demotivational quotes}
\format{
a data frame with 23 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
tv
}
\description{
This is a data frame of tv and film demotivational quotes taken from American film and TV shows.
}
\keyword{datasets}
