% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{work}
\alias{work}
\title{A data frame of work-related demotivational quotes}
\format{
a data frame with 31 observations and 1 column
\describe{
\item{\code{values}}{a character vector}
}
}
\usage{
work
}
\description{
This is a data frame of work-related demotivational quotes
}
\keyword{datasets}
